actions.py aufgeteilt in einzelne Bereiche
