+++ actions/actions.py
@@ -1,5 +1,6 @@
 import collections
 import logging
+
 from typing import Any, Optional, Text, Dict, List
 from rasa_sdk.executor import CollectingDispatcher
 
@@ -11,309 +12,6 @@ from rasa_sdk import Action, Tracker
 
 
 logger = logging.getLogger(__name__)
-
-class ActionProcessInformAll(Action):
-    def name(self) -> Text:
-        return "process_inform_all"
-    
-    def run(
-        self, 
-        dispatcher: "CollectingDispatcher", 
-        tracker: Tracker, 
-        domain: "DomainDict"
-    ) -> List[Dict[Text, Any]]:
-        drug_slot = tracker.get_slot("drogen")
-        no_drug_slot = tracker.get_slot("no_drogen")
-        pet_slot = tracker.get_slot("haustier")
-        no_pet_slot = tracker.get_slot("no_haustier")
-        shelter_slot = tracker.get_slot("unterkunft")
-        apartment_slot = tracker.get_slot("wohnung")
-
-        petowner = SlotSet("chaustierhalter", None)
-        drug_addict = SlotSet("drogenabhaengig", None)
-        shelter = SlotSet("unterkunft", None)
-
-        if drug_slot is not None:
-            if no_drug_slot is None:
-                drug_addict = SlotSet("drogenabhaengig", True)
-        else:
-            if no_drug_slot is not None:
-                drug_addict = SlotSet("drogenabhaengig", False)
-
-        if pet_slot is not None:
-            if no_pet_slot is None:
-                petowner = SlotSet("chaustierhalter", True)
-        else:
-            if no_pet_slot is not None:
-                petowner = SlotSet("chaustierhalter", False)
-
-        
-        if shelter_slot is not None:
-            if apartment_slot is None:
-                shelter = SlotSet("notunterkunft", True)
-        else:
-            if apartment_slot is not None:
-                shelter = SlotSet("notunterkunft", False)
-        
-        return [petowner, drug_addict, shelter]
-
-class ValidateMedicineForm(FormValidationAction):
-
-    def name(self) -> Text:
-        return "validate_medicine_form"
-
-    def validate_emergency(
-        self,
-        slot_value: bool,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-
-        # zeigt beim nächsten mal slot_value-type an
-        text = f">>>>Slot emergency ist gesetzt auf '{slot_value}' und vom Typ {type(slot_value)}"
-
-        intent = tracker.latest_message['intent'].get('name')
-
-        if(intent == 'affirm'):
-            return{"emergency": True}
-        elif(intent == 'deny'):
-            return{"emergency": False}
-        elif(intent == 'medicine_specific_emergency' or intent == 'medicine_all_information'):
-            return {"emergency": slot_value}
-        else:
-            dispatcher.utter_message(
-                text="Entschuldigung, die Eingabe zu >>Notfall<< wurde nicht erkannt. Du kannst auf diese Frage mit 'Ja' oder 'Nein' antworten.")
-            return {"emergency": None}
-
-    def validate_insurancestatus(
-            self,
-            slot_value: bool,
-            dispatcher: CollectingDispatcher,
-            tracker: Tracker,
-            domain: DomainDict,) -> Dict[Text, Any]:
-
-        intent = tracker.latest_message['intent'].get('name')
-
-        if(intent == 'affirm'):
-            return{"insurancestatus": True}
-        elif(intent == 'deny'):
-            return{"insurancestatus": False}
-        elif(intent == 'medicine_specific_insurancestatus' or intent == 'medicine_all_information'):
-            return {"insurancestatus": slot_value}
-        else:
-            dispatcher.utter_message(
-                text="Entschuldigung, die Eingabe zu >>Versicherungsstatus<< wurde nicht erkannt. Du kannst auf diese Frage mit 'Ja' oder 'Nein' antworten.")
-            return {"insurancestatus": None}
-
-
-class ValidateInformation(FormValidationAction):
-    def name(self) -> Text:
-        return "validate_informationen_form"
-
-    def validate_alter(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        try:
-            if int(slot_value) >= 18:
-                if int(slot_value) >= 22:
-                    return {"volljaehrig": True, "jugendlich": False}
-                else:
-                    return {"volljaehrig": True, "jugendlich": True}
-            else:
-                return {"volljaehrig": False, "jugendlich": True}
-        except:
-            dispatcher.utter_message(
-                text="Die Eingabe zu >>Alter<< wurde nicht erkannt. Bitte gib eine Zahl an, die dein >>Alter<< darstellt.")
-            return {"alter": None}
-    
-    def validate_bgeschlecht(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if slot_value.lower() in ["männlich", "weiblich", "divers"]:
-            return {"bgeschlecht": slot_value.lower()}
-        else:
-            dispatcher.utter_message(text="Es tut mir Leid, die Eingabe für dein >>Geschlecht<< wurde nicht erkannt. Bitte gib an, ob du dich als 'männlich', 'weiblich' oder 'divers' bezeichnen würdest.")
-            return {"bgeschlecht": None}
-
-    def validate_chaustierhalter(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if type(slot_value) is bool:
-            return {"chaustierhalter": slot_value}
-        else:
-            dispatcher.utter_message(text="Entschuldigung, die Eingabe zu >>Haustieren<< wurde nicht erkannt. Du kannst auf diese Frage mit 'Ja' oder 'Nein' antworten.")
-            return {"chaustierhalter": None}
-
-    def validate_drogenabhaengig(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if type(slot_value) is bool:
-            return {"drogenabhaengig": slot_value}
-        else:
-            dispatcher.utter_message(
-                text="Es tut mir Leid, die Eingabe für >>Drogen<< wurde nicht erkannt. Du kannst auf eine Frage zu >>Drogen<< mit 'Ja' oder 'Nein' antworten.")
-            return {"drogenabhaengig": None}
-
-
-class ValidateNotunterkunft(FormValidationAction):
-    def name(self):
-        return "validate_unterkunft_form"
-
-    def validate_notunterkunft(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if slot_value is not None and type(slot_value) is bool:
-            return {"notunterkunft": slot_value}
-        else:
-            dispatcher.utter_message(
-                text="Entschuldigung, die Eingabe für >>Notunterkunft<< wurde nicht erkannt. Du kannst auf diese Frage mit 'Ja' oder 'Nein' antworten.")
-            return {"notunterkunft": None}
-
-
-class ValidateHygieneartikel(FormValidationAction):
-    def name(self):
-        return "validate_hygiene_form"
-    
-    def validate_hygieneartikel(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if type(slot_value) is list:
-            return {"hygieneartikel": slot_value}
-        else:
-            dispatcher.utter_message(
-                text="Es tut mir Leid, aber es scheint nicht geklappt zu haben. Du kannst hier einfach alle gewünschten >>Hygieneartikel<< auflisten.")
-            return {"hygieneartikel": None}
-
-class ActionAnswerHygiene(Action):
-    def name(self):
-        return "action_answer_hygiene"
-
-    def run(
-        self,
-        dispatcher: "CollectingDispatcher",
-        tracker: Tracker,
-        domain: "DomainDict"
-    ) -> List[Dict[Text, Any]]:
-        answer = "Alles klar. Diese Artikel:\n"
-
-        slot_value = tracker.get_slot("hygieneartikel")
-
-        for item in slot_value:
-            answer += f"  - {item}\n"
-
-        answer += f"\nFindest du bei diesen Stellen:"
-        answer += f"\n - mudra"
-        answer += f"\n - Obdachlosenhilfe Nürnberg"
-        answer += f"\n - Heinzelmännchen"
-        answer += f"\n - Nürnberger Engel"
-
-        dispatcher.utter_message(text=answer)
-
-        return []
-
-class ValidateSpendenartikel(FormValidationAction):
-    def name(self):
-        return "validate_spendenartikel_form"
-    
-    def validate_spendenartikel(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if type(slot_value) is list:
-            return {"spendenartikel": slot_value}
-        else:
-            dispatcher.utter_message(text="Es tut mir Leid, aber es scheint nicht geklappt zu haben. Du kannst hier einfach alle gewünschten >>Spendenartikel<< auflisten.")
-            return {"spendenartikel": None}
-
-class ActionAnswerSpende(Action):
-    def name(self):
-        return "action_answer_spende"
-
-    def run(
-        self, 
-        dispatcher: "CollectingDispatcher", 
-        tracker: Tracker, 
-        domain: "DomainDict"
-    ) -> List[Dict[Text, Any]]:
-        answer = "";
-
-        hygiene_slot = tracker.get_slot("hygieneartikel")
-        spendenartikel = tracker.get_slot("spendenartikel")
-        spendeErhalten = tracker.get_slot("spendeErhalten")
-
-        spende = list()
-        spende_schlafen = list()
-
-        if spendenartikel:
-            for item in spendenartikel:
-                temp = item.lower()
-                if temp == "decke" or temp == "schlafsack" or temp == "isomatte" or temp == "kissen" :
-                    spende_schlafen.append(item)
-                else:
-                    spende.append(item)
-        
-        if len(spende_schlafen) != 0:
-            answer += "Diese Artikel:\n"
-            for item in spende_schlafen:
-                answer += f"  - {item}\n"
-            if spendeErhalten:
-                answer += f"\nFindest du bei diesen Stellen: Obdachlosenhilfe"
-            else: 
-                answer += f"\nKannst du bei diesen Stellen abgeben: Obdachlosenhilfe"
-
-        if len(spende) != 0:
-            answer += "\nDiese Artikel:\n"
-            for item in spende:
-                answer += f"  - {item}\n"
-            if spendeErhalten:
-                answer += f"\nFindest du bei diesen Stellen: (noch nicht im Prototyp hinterlegt)"
-            else: 
-                answer += f"\nKannst du bei diesen Stellen abgeben: (noch nicht im Prototyp hinterlegt)"
-
-        if hygiene_slot and len(hygiene_slot) != 0:
-            answer += "\nDie Hygieneartikel:\n"
-            for item in hygiene_slot:
-                answer += f"  - {item}\n"
-            if spendeErhalten:
-                answer += f"\nFindest du bei diesen Stellen:"
-                answer += f"\n - mudra"
-                answer += f"\n - Obdachlosenhilfe Nürnberg"
-                answer += f"\n - Heinzelmännchen"
-                answer += f"\n - Nürnberger Engel"
-            else:
-                answer += f"\nkönnen nicht gespendet werden."
-
-        dispatcher.utter_message(text=answer)
-
-        return []
     
 class ActionAnswerURL(Action):
     def name(self):
@@ -358,20 +56,4 @@ class ActionAnswerURL(Action):
 
         return []
 
-        
-class ValidateSpendeErhalten(FormValidationAction):
-    def name(self):
-        return "validate_spendeErhalten_form"
 
-    def validate_spendeErhalten(
-        self,
-        slot_value: Any,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: DomainDict
-    ) -> Dict[Text, Any]:
-        if slot_value is not None and type(slot_value) is bool:
-            return {"spendeErhalten": slot_value}
-        else:
-            dispatcher.utter_message(text="Entschuldigung, die Eingabe für >>spendeErhalten<< wurde nicht erkannt. Du kannst auf diese Frage mit 'Ja' oder 'Nein' antworten.")
-            return {"spendeErhalten": None}
